Create a function that returns multiple strings
In the first lesson, learned the beginnings of testing javascript with
the terminal.

The second lesson required a loop to add the string to an empty
variable. Went to discord and got help where others had similar issues.

At this time it is helpful to think of parameters as variables I can
change to return a different result.